Change-schema-type

+++ src/feature/books/schema.ts
@@ -1,9 +1,9 @@
-import { integer, pgTable, varchar } from "drizzle-orm/pg-core";
+import { integer, pgTable, varchar, numeric } from "drizzle-orm/pg-core";
 
 export const booksTable = pgTable("books", {
   id: integer().primaryKey().generatedAlwaysAsIdentity(),
   title: varchar({ length: 120 }).notNull(),
   description: varchar({ length: 250 }).notNull(),
-  price: integer().notNull(),
+  price: numeric().notNull(), // ÅTERSKAPA FEL SÅ ATT VI KAN FÅNGA DET!!!!
   author_id: integer().notNull(),
 });
